Remove duplicate `__pattern_min_element_reduce_fn` implementations (#2205)
